LIBHYDRA-260. Moved Plastron status enumeration into separate Concern
Created "PlastronStatus" concern to hold the "plastron_status" enum
used by ExportJob and ImportJob. This removes the code duplication
from these classes.

https://issues.umd.edu/browse/LIBHYDRA-260
